requirements.txt
streamlit
sentence-transformers
faiss-cpu
openai
python-dotenv